parts: LDraw 2008-01 update.
This commit is the result of running LDraw0801.exe.

The following installation options are selected:
- Core parts library
- Alias parts library
- Physical colour parts library

The following option is NOT selected:
- Non re-distributable parts library

This also adds models/note0801*.txt files which were added to subsequent
updates for backward compatibility.

Official release notes, quoted from ldraw.org:

"""
LDraw.org Library Update 2008-01 September 10, 2008
This release is complete re-issue of the official library, up to and including the last (2005-01) parts update, with the primary purpose of aligning the library with the Contributor Agreement. This clarifies the redistribution status of the part files within the library, and standardises the file headers.

To this end, a total of 90 files have been excluded from the re-distributable library because we have been unable to contact the original author or persuade them to affirm the Contributor Agreement. The current community of LDraw part authors is working to re-create those parts so that they can be added to the re-distributable library in a subsequent parts update.

In addition to separating the parts that are not re-distributable, We have taken the opportunity to separate two additional classes of part files and make their installation optional in a new Windows installation package.

- Alias parts : generic colour versions of parts that are physically identical to parts in the core parts library, but have a different part number, either because of production differences between opaque and transparent parts or due to evolution of the part numbering scheme.

- Physical colour parts : hard-coded colour versions of parts or composite parts.

No new parts are included in this update, but its release does now make it possible to restart the more regular publication of additions to the library. Minor changes have been made to part descriptions as part of the header standardisation.
"""