parts: Include unofficial lsynth parts.
This includes the required parts to view LDraw files
generated by lsynth.

Including it in the parts folder ensures that all LDraw tools
can find them without special configuration. Since this is Git,
there is no penalty in doing this, because this commit can
just be reverted if needed.

This is mainly for convenience, so that it is no longer required to
search the web for three different zip archives and unpack them in
specific locations just to make lsynth work.